fix(integer-i64): widen parallel worker extra-arg push from i32 to i64
In `State::execute_at_raw` (the worker-thread entrypoint for
parallel_for / parallel_for_light), each extra context argument
was pushed onto the worker's stack as `i32` (4 bytes) before
invoking the worker function.  Post-2c loft `integer` is i64
(8 bytes), and the worker function reads its parameters as i64
— the size mismatch caused every subsequent stack read to land
one slot-offset too low.

Three failure modes all fell out of this single bug:

- **Category E** `par_light_extra_args`: worker panicked with
  "No elements left on the stack 8 < 8" when the function body
  tried to read past the misaligned stack top.
- **Category D.2** `wrap::script_threading`: `codegen.rs:1780`
  compile-time assertion `var b[520] versus 517` — the tracker
  and runtime stacks agreed at compile time but drifted at
  runtime because workers pushed 4 B where the caller expected
  8 B.
- **Category D.3** `wrap::threading`: `ops.rs:278` long
  overflow `i64::MIN | u32::MAX + -2^32` — the worker read a
  garbage i64 whose low 32 bits came from its neighbouring
  slot (a genuine null sentinel's high half).

Change: `put_stack(extra as i32)` → `put_stack(extra as i64)`.
Symmetric with the codegen fix in `7bf3558` which widened the
`n_parallel_for` extra-arg cleanup tracker from `-= 4` to
`-= 8`.

Tests: 19 → 16 failing.  Cleared:
- `par_light_extra_args` (Category E — closes the cluster)
- `wrap::script_threading` (Category D.2 — closes the cluster)
- `wrap::threading` (Category D.3 — closes the cluster)

+++ src/state/mod.rs
@@ -1727,8 +1727,8 @@ impl State {
         // Element arg (DbRef) occupies the first parameter slot; extras follow.
         self.put_stack(*arg); // 12 bytes
         for &extra in extra_args {
-            // Push each extra as a raw i32 (integer context args).
-            self.put_stack(extra as i32);
+            // Push each extra as a raw i64 (integer context args, post-2c).
+            self.put_stack(extra as i64);
         }
         self.put_stack(u32::MAX); // return address sentinel
         self.code_pos = fn_pos;
